requirements.txt
mysql-connector-python
flask
Flask-WTF
